revise

+++ muft/server/controllers/controllers.js
@@ -33,26 +33,34 @@ export const getStation = (req, res) => {
 };
 
 export const updateStation = (req, res) => {
-  let { id } = req.params;
-  Station.findOneAndUpdate(
-    { _id: mongoose.Types.ObjectId(id) },
-    req.body,
-    { new: true, upsert: true },
-    (err, Station) => {
-      if (err) {
-        res.send(err);
+  let { id, passKey } = req.params;
+  if (passKey === process.env.PASS_KEY) {
+    Station.findOneAndUpdate(
+      { _id: mongoose.Types.ObjectId(id) },
+      req.body,
+      { new: true, upsert: true },
+      (err, Station) => {
+        if (err) {
+          res.send(err);
+        }
+        res.json(Station);
       }
-      res.json(Station);
-    }
-  );
+    );
+  } else {
+    res.json({ error: `Pass key is not valid. Please try again.` });
+  }
 };
 
 export const deleteStation = (req, res) => {
-  let { id } = req.params;
-  Station.deleteOne({ _id: mongoose.Types.ObjectId(id) }, (err, Station) => {
-    if (err) {
-      res.send(err);
-    }
-    res.json({ message: `Station deleted!` });
-  });
+  let { id, passKey } = req.params;
+  if (passKey === process.env.PASS_KEY) {
+    Station.deleteOne({ _id: mongoose.Types.ObjectId(id) }, (err, Station) => {
+      if (err) {
+        res.send(err);
+      }
+      res.json({ message: `Station deleted!` });
+    });
+  } else {
+    res.json({ error: `Pass key is not valid. Please try again.` });
+  }
 };
